requirements.txt
# uv venv venv --python 3.9.22
# venv\Scripts\activate
# uv pip install -r requirements.txt
numpy==1.19.5 # From 1.17.4
matplotlib==3.3.4 # From 3.1.2
scipy==1.5.4 # From 1.3.3
openpyxl>=3.0.5
wxPython==4.1.1
pyserial>=3.4
black
mypy
isort